Modified Dash functionality, added Double Jump
Character dashes directly forward by default - that is, if the character is standing still. If a direction is being held, the character will dash in that direction.
Dashing halts all vertical aerial momentum. Character can now double jump by pressing the Space key or "A" face button in midair. Dashing can be canceled into a double jump, but lateral momentum will be preserved.